refactor: コンポーネントに切り出した(元のメッセージが削除された際にメッセージが出るようにも変更)
