+++ web/src/features/message/type.ts
@@ -3,6 +3,6 @@ export type Message = {
   author: string;
   content: string;
   is_me: boolean;
-  icon: string;
-  reply_to_id: string | undefined;
+  icon: string | null;
+  reply_to_id: string | null;
 };
