+++ web/src/features/routes/chat/messagecontainer.module.scss
@@ -12,7 +12,6 @@
     padding: 1em;
     margin: 0.5em 0;
 
-
     &.is_me {
       flex-direction: row-reverse;
       padding-left: 10em;
@@ -46,11 +45,13 @@
     .message {
       box-sizing: border-box;
       padding: 0.7em 1em;
-      white-space: pre-line;
       line-break: anywhere;
 
       .message_reply_area {
-        display: flex;
+        height: auto;
+        text-align: left;
+        background: transparent;
+        border-color: black;
       }
     }
   }
